Major changes. AuthController is implemented for the authentication and registration operations, GlobalExceptionHandler has been added, JWT token logic is chosen and implemented

+++ src/main/java/com/example/soapz/Models/ServiceRegistry/ServiceCategory.java
@@ -6,13 +6,13 @@ import lombok.*;
 
 @Entity
 @Table(name = "service_category")
-@Data // Включает геттеры, сеттеры, toString, equals, hashCode
+@Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ServiceCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "name", nullable = false, length = 100)
     private String name;
